move shared parser function to utils
