+++ src/zampy/datasets/era5_land.py
@@ -14,7 +14,6 @@ from zampy.datasets.dataset_protocol import TimeBounds
 from zampy.datasets.dataset_protocol import Variable
 from zampy.datasets.dataset_protocol import copy_properties_file
 from zampy.datasets.dataset_protocol import write_properties_file
-from zampy.reference.variables import VARIABLE_REFERENCE_LOOKUP
 from zampy.reference.variables import unit_registry
 from zampy.utils import regrid
 
@@ -101,7 +100,7 @@ class ERA5Land(Dataset):  # noqa: D101
         data_files = list(download_folder.glob(data_file_pattern))
 
         for file in data_files:
-            convert_to_zampy(
+            utils.convert_to_zampy(
                 ingest_folder,
                 file=file,
                 overwrite=overwrite,
@@ -153,65 +152,3 @@ class ERA5Land(Dataset):  # noqa: D101
             # call ds.compute()
 
         return True
-
-
-def convert_to_zampy(
-    ingest_folder: Path,
-    file: Path,
-    overwrite: bool = False,
-) -> None:
-    """Convert the downloaded nc files to standard CF/Zampy netCDF files.
-
-    The downloaded ERA5-land data already follows CF1.6 convention. However, it uses
-    (abbreviated) variable name instead of standard name, which prohibits the format
-    conversion. Therefore we need to ingest the downloaded files and rename all
-    variables to standard names.
-
-    Args:
-        ingest_folder: Folder where the files have to be written to.
-        file: Path to the ERA5-land nc file.
-        overwrite: Overwrite all existing files. If False, file that already exist will
-            be skipped.
-    """
-    ncfile = ingest_folder / file.with_suffix(".nc").name
-    if ncfile.exists() and not overwrite:
-        print(f"File '{ncfile.name}' already exists, skipping...")
-    else:
-        ds = parse_nc_file(file)
-
-        ds.to_netcdf(path=ncfile)
-
-
-var_reference_era5_land_to_zampy = {
-    "t2m": "air_temperature",
-    "strd": "dewpoint_temperature",
-}
-
-
-def parse_nc_file(file: Path) -> xr.Dataset:
-    """Parse the downloaded ERA5 nc files, to CF/Zampy standard dataset.
-
-    Args:
-        file: Path to the ERA5 nc file.
-
-    Returns:
-        CF/Zampy formatted xarray Dataset
-    """
-    # Open chunked: will be dask array -> file writing can be parallelized.
-    ds = xr.open_dataset(file, chunks={"x": 50, "y": 50})
-
-    for variable in ds.variables:
-        if variable in var_reference_era5_land_to_zampy:
-            var = str(variable)  # Cast to string to please mypy
-            variable_name = var_reference_era5_land_to_zampy[var]
-            ds = ds.rename({var: variable_name})
-            ds[variable_name].attrs["units"] = str(
-                VARIABLE_REFERENCE_LOOKUP[variable_name].unit
-            )
-            ds[variable_name].attrs["description"] = VARIABLE_REFERENCE_LOOKUP[
-                variable_name
-            ].desc
-
-    # TODO: add dataset attributes.
-
-    return ds
